requirements.txt
alembic==1.16.1
asttokens==3.0.0
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dotenv==0.9.9
executing==2.2.0
fonttools==4.58.2
greenlet==3.2.3
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.29.5
ipython==9.3.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
joblib==1.5.1
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyterlab-widgets==3.0.15
kds==0.1.3
kiwisolver==1.4.8
mako==1.3.10
markupsafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.0.2
nvidia-nccl-cu12==2.27.3
optuna==4.2.1
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt-toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
sqlalchemy==2.0.41
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.14.0
tzdata==2025.2
wcwidth==0.2.13
widgetsnbextension==4.0.14
xgboost==3.0.0
